fix(#3304): update landing.vue and default.vue layout files
