+++ packages/docs/layouts/default.vue
@@ -23,8 +23,12 @@
 </template>
 
 <script setup lang="ts">
+import { useColors } from 'vuestic-ui'
 import { useDocsScroll } from '../composables/useDocsScroll';
 
+const colorMode = useColorMode()
+const cookie = useCookie('vuestic-theme')
+const { applyPreset } = useColors()
 const breakpoints = useBreakpoint()
 
 const isSidebarVisible = ref(!breakpoints.smDown)
@@ -47,19 +51,22 @@ afterEach(() => {
     isSidebarVisible.value = false
   }
 })
+
+onBeforeMount(() => {
+  applyPreset(cookie.value || colorMode.preference)
+})
+
 onMounted(scrollToElement)
 
 useHead({
-  meta: {
-    link: [
-      { href: 'https://cdn.jsdelivr.net/npm/@mdi/font@5.9.55/css/materialdesignicons.min.css', rel: 'stylesheet' },
-      { href: 'https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css', rel: 'stylesheet' },
-    ],
-    script: [
-      { src: 'https://kit.fontawesome.com/5460c87b2a.js', crossorigin: 'anonymous' },
-      { src: 'https://unpkg.com/ionicons@5.5.2/dist/ionicons/ionicons.esm.js', type: 'module' },
-    ],
-  } as any
+  link: [
+    { href: 'https://cdn.jsdelivr.net/npm/@mdi/font@5.9.55/css/materialdesignicons.min.css', rel: 'stylesheet' },
+    { href: 'https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css', rel: 'stylesheet' },
+  ],
+  script: [
+    { src: 'https://kit.fontawesome.com/5460c87b2a.js', crossorigin: 'anonymous' },
+    { src: 'https://unpkg.com/ionicons@5.5.2/dist/ionicons/ionicons.esm.js', type: 'module' },
+  ],
 })
 </script>
